# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sentinel Global Situational Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetAffiliation } from "./assetAffiliation";
import type { AssetStatus } from "./assetStatus";
import type { AssetType } from "./assetType";

export interface Asset {
  id: number;
  name: string;
  designation?: string;
  type: AssetType;
  status: AssetStatus;
  affiliation: AssetAffiliation;
  country?: string;
  latitude: number;
  longitude: number;
  altitude?: number | null;
  speed?: number | null;
  /** Bearing in degrees 0-360 */
  heading?: number | null;
  trackId?: string | null;
  lastPositionAt: Date;
  updatedAt: Date;
}