# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sentinel Global Situational Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummarySystemStatus } from "./dashboardSummarySystemStatus";

export interface DashboardSummary {
  totalThreats: number;
  criticalThreats: number;
  activeIncidents: number;
  gpsAnomalies: number;
  trackedAssets: number;
  activeSignals: number;
  /** Change in threat count vs 24h ago */
  threatDelta24h: number;
  systemStatus: DashboardSummarySystemStatus;
  lastUpdated: Date;
}