# lib/api-zod/src/generated/types/threatTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sentinel Global Situational Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ThreatTrendPoint {
  hour: string;
  critical: number;
  high: number;
  medium: number;
  low: number;
  total: number;
}